faq/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1
Werkzeug==2.3.7
Flask-CORS==4.0.0
requests==2.31.0
